Send memory usage information together with metrics on Linux
We handle only Linux since getting accurate memory usage values on it is
most important. Memory cannot be read platform independently in Python
and we don't want at this point to include any external dependencies,
such as `psutil`. We might want to add support for other systems when
the need arises for some user, it likely won't happen soon since since
production web servers are not often ran on non-Linux systems.

Good to note that `MemAvailable` in `/proc/meminfo` is only available on
Linux kernel 3.14 and up, but since 3.14 has already reached end of
life we can pretty safely ignore support for systems before it.